Minor fixes, update About page - Clean code a tiny bit - Add confirmation to news post deletion

+++ src/components/news/NewsPost.vue
@@ -58,6 +58,9 @@ export default Vue.extend({
 
   methods: {
     async handleDeleteNewsPost() {
+      if (!window.confirm('Are you sure you want to delete this news post?')) {
+        return;
+      }
       try {
         await deleteNewsPost(this.id);
         this.displayStatus(
